Leave K loop in reedSolomon::computeLambda in time
So far, the loop ran as long as K <= nroots. In the last iteration, the
only relevant operation was the update of Lambda. All operations after
that have no relevance if there is no next iteration. However, that part
caused an invalid access into syndromes.

Let the loop run only as long K < nroots and do the final Lambda
calculation afterwards.

Fixes #202