hackrf: update gain display when restoring settings
So far, only the sliders are updated and not the LCD type numeric
displays. Since signal emission is disabled during restore, the updates
do not automatically propagate to the numeric displays.

This commit also removes leftover definitions of __KEEP_GAIN_SETTINGS__
in qt-dab.pro since this setting is gone since qt-dab 3.5.